- Simplify some conditions - Fix last migration from `DynaicVectorClass` causing crash - Fix `GainSelfHealing` check was not written properly causing it to broke

+++ src/Ext/Techno/Hooks.Firing.FLH.cpp
@@ -17,7 +17,7 @@ DEFINE_HOOK(0x6F3AF9, TechnoClass_GetFLH_GetAlternateFLH, 0x5)
 	const auto conpy_weaponIdx_B = abs(5 + weaponIdx);
 	Debug::Log("[%s] Trying to get Additional AlternateFLH at [original %d vs changed %d] !\n", pTypeExt->AttachedToObject->ID, conpy_weaponIdx , conpy_weaponIdx_B);
 
-	if(conpy_weaponIdx_B < static_cast<int>(pTypeExt->AlternateFLHs.size())) {
+	if((size_t)conpy_weaponIdx_B < pTypeExt->AlternateFLHs.size()) {
 		const CoordStruct& flh = pTypeExt->AlternateFLHs[conpy_weaponIdx_B] ;
 
 		R->ECX(flh.X);
